+++ spec/factory_sloth/cli_spec.rb
@@ -1,12 +1,19 @@
 describe FactorySloth::CLI, '::call' do
+  let(:result_stub) { FactorySloth::CodeMod.new('dummy_code') }
+  before { allow(result_stub).to receive(:create_calls).and_return([])}
+  before { allow(result_stub).to receive(:changed_create_calls).and_return([])}
+
   it 'takes paths as arguments' do
     input = fixture('build_ok')
     temp_path = "#{Dir.tmpdir}/build_ok"
     File.write(temp_path, input)
 
-    expect { FactorySloth::CLI.call([temp_path]) }.to output(
-      /Processing .*build_ok.*- create in line 3.* 1 create calls found, 1 replaced/m
-    ).to_stdout
+    expect { FactorySloth::CLI.call([temp_path]) }.to output(Regexp.new([
+      /Processing .*build_ok.*/,
+      /- create in line 3 can be replaced with build.*/,
+      /.* 1 create calls found, 1 replaced.*/,
+      /Scanned 1 files, found 1 unnecessary create calls across 1 files/,
+    ].join("\n+"))).to_stdout
     result = File.read(temp_path)
     expect(result).not_to eq input
     expect(result).to include 'build(:optional_create)'
@@ -16,21 +23,29 @@ describe FactorySloth::CLI, '::call' do
 
   it 'forces processing when given individual files, even if done previously' do
     expect(FactorySloth::FileProcessor).to receive(:process)
-    expect { FactorySloth::CLI.call([__FILE__]) }.to output.to_stdout
+      .and_return(result_stub)
+    expect { FactorySloth::CLI.call([__FILE__]) }
+      .to output(/Scanned 1 files.* found 0 unnecessary create calls/).to_stdout
 
     expect(FactorySloth::FileProcessor).to receive(:process)
-    expect { FactorySloth::CLI.call([__FILE__]) }.to output.to_stdout
+      .and_return(result_stub)
+    expect { FactorySloth::CLI.call([__FILE__]) }
+      .to output(/Scanned 1 files.* found 0 unnecessary create calls/).to_stdout
   end
 
   it 'can force processing with the --force option' do
     n = Dir["#{__dir__}/**/*_spec.rb"].count
     expect(FactorySloth::FileProcessor).to receive(:process).exactly(n).times
-    expect { FactorySloth::CLI.call([__dir__]) }.to output.to_stdout
+      .and_return(result_stub)
+    expect { FactorySloth::CLI.call([__dir__]) }
+      .to output(/Scanned #{n} files/).to_stdout
 
     expect(FactorySloth::DoneTracker.done?(__FILE__)).to eq true
 
     expect(FactorySloth::FileProcessor).to receive(:process).exactly(n).times
-    expect { FactorySloth::CLI.call([__dir__, '--force']) }.to output.to_stdout
+      .and_return(result_stub)
+    expect { FactorySloth::CLI.call([__dir__, '--force']) }
+      .to output(/Scanned #{n} files/).to_stdout
   end
 
   it 'can lint with the --lint option' do
@@ -40,7 +55,7 @@ describe FactorySloth::CLI, '::call' do
     expect { FactorySloth::CLI.call(["#{__dir__}/../fixtures/build_ok.rb", '--lint']) }
       .to raise_error(SystemExit)
       .and output(/1 create calls found, 1 replaceable/).to_stdout
-      .and output(/build_ok\.rb/).to_stderr
+      .and output(/1 unnecessary create calls across 1 files:\n.*build_ok\.rb/).to_stderr
   end
 
   it 'can output help and version' do